# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Logtailr API",
    "description": "REST API for the Logtailr concurrent multi-source log aggregator. Provides health monitoring, source status, configuration inspection, Prometheus metrics, and real-time log streaming via WebSocket.",
    "version": "0.5.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server (default --api-port)"
    }
  ],
  "paths": {
...
    "/ws/logs": {
      "get": {
        "operationId": "streamLogs",
        "summary": "WebSocket log stream",
        "description": "Upgrades to a WebSocket connection that streams log lines in real-time as JSON objects. Supports optional query parameter filters for level and source.",
        "tags": ["WebSocket"],
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "description": "Minimum log level filter (debug, info, warn, error, fatal). Only logs at this level or above are sent.",
            "schema": {
              "type": "string",
              "enum": ["debug", "info", "warn", "error", "fatal"]
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "description": "Filter by source name. Only logs from this source are sent.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "101": {
            "description": "WebSocket upgrade successful. Each message is a JSON-encoded LogLine."
          }
        }
      }
    }
  },
...
}